+++ components/config-provider/config-provider.ts
@@ -1,5 +1,5 @@
-import { defineComponent, toRef, renderSlot } from 'vue'
-import { configProps, configLocale, configIcons } from '@vexip-ui/config'
+import { defineComponent, toRefs, renderSlot } from 'vue'
+import { configProps, configLocale, configIcons, configZIndex } from '@vexip-ui/config'
 
 import type { PropType } from 'vue'
 import type { LocaleOptions, IconsOptions } from '@vexip-ui/config'
@@ -19,12 +19,19 @@ export default defineComponent({
     icons: {
       type: Object as PropType<IconsOptions>,
       default: () => ({})
+    },
+    zIndex: {
+      type: Number,
+      default: null
     }
   },
-  setup(props, { slots }) {
-    configProps(toRef(props, 'props'))
-    configLocale(toRef(props, 'locale'))
-    configIcons(toRef(props, 'icons'))
+  setup(_props, { slots }) {
+    const { props, locale, icons, zIndex } = toRefs(_props)
+
+    configProps(props)
+    configLocale(locale)
+    configIcons(icons)
+    configZIndex(zIndex)
 
     return () => renderSlot(slots, 'default')
   }