+++ docs/demos/native-scroll/pointer/demo.zh-CN.vue
@@ -1,0 +1,15 @@
+<template>
+  <div style="max-width: 360px; border: var(--vxp-border-base)">
+    <NativeScroll :height="120" pointer>
+      <p v-for="n in 9" :key="n">
+        {{ n }}、在此处滚动鼠标滚轮
+      </p>
+    </NativeScroll>
+  </div>
+</template>
+
+<style scoped>
+.vxp-native-scroll p {
+  padding: 5px 10px;
+}
+</style>